move lock code to file_operation
